Add `--inherit-stdin` as a command line option
Add a new option for the program launched by alacritty to inherit
alacritty's stdin file descriptor.  A contrived example for using this
option is something like `dmesg | alacritty --inherit-stdin -e vim`,
where vim will be able to read the output from `dmesg`.

In general users should not need to use this directly.  The main user is
expected to be alacritty itself where it will spawn a new instance that
runs a pager and then dump the contents of the scrollback buffer into
the stdin pipe of the new instance.